Rename files (and update imports) to enforce naming convention on config components
