+++ web/src/pages/Settings/Config/ConfigPanels/UploadConfig.jsx
@@ -3,17 +3,17 @@ import React from 'react';
 import "../Config.scss";
 import ConfigPanel from './ConfigTemplates/ConfigPanel.jsx';
 import Button from '@mui/material/Button/Button';
-import ConfigModal from './ConfigTemplates/ConfigModal';
+import ConfigModal from './ConfigTemplates/ConfigModal.jsx';
 import StatusBar from './ConfigTemplates/StatusBar.jsx';
 
-export default function ConfigDownload(){
+export default function UploadConfig(){
     const [file, setFile] = React.useState([]);
     const [filePicked, setFilePicked] = React.useState(false);
     const [modalOpen, setModalOpen] = React.useState(false);
     const [loading, setLoading] = React.useState(false);
     const [response, setResponse] = React.useState(null);
 
-    async function UploadConfig(){
+    async function ConfigUpload(){
         setLoading(true);
         const resp = await fetch("/api/load", {
             method: "POST",
@@ -54,7 +54,7 @@ export default function ConfigDownload(){
 
             <ConfigModal
                 body={"This will replace the previous config, ensure you've downloaded the current config if you wish to keep it!"}
-                confirm={() => {UploadConfig();}}
+                confirm={() => {ConfigUpload();}}
                 open={modalOpen}
                 setOpen={setModalOpen}
             />
